+++ src/main/java/com/example/workus/approval/dto/ApprovalForm.java
@@ -1,0 +1,8 @@
+package com.example.workus.approval.dto;
+
+
+public class ApprovalForm {
+
+    private String title;
+    private String text;
+}
