+++ src/main/java/com/example/workus/attendance/controller/RestAttendanceController.java
@@ -62,5 +62,4 @@ public class RestAttendanceController {
         return ResponseEntity.ok(Map.of("events", events, "depts", depts, "roleNo", roleNo));
     }
 
-
 }
