+++ src/main/java/com/example/workus/user/mapper/UserMapper.java
@@ -32,8 +32,12 @@ public interface UserMapper {
     void updateMyUser(@Param("user") User user); // 내 정보를 수정한다.
 
     List<DeptDto> getDeptByUserNo(@Param("userNo") long userNo);
-    
+
     void updateMyPhone(@Param("user") User user); // 내 연락처를 수정한다.
-    
-    void updateMyPassword(@Param("user") User user); // 내 비밀번호를 수정한다.
+
+    void updateMyPassword(@Param("user") User user); // 내 비밀번호를 수정한다
+
+    int getUserRoleNo(@Param("userNo") Long userNo); // 권한 조회
+
+    User getLeaderByDeptNo(@Param("deptNo") long deptNo); // 팀장 조회
 }
