+++ src/main/webapp/resources/css/attendance.css
@@ -9,15 +9,15 @@
 .lnb li a:after {
     content: "";
     left: 0;
-    bottom: 0;
+    bottom: 3px;
     width: 0;
-    height: 2px;
+    height: 9px;
     z-index: -1;
     background-color: #e91e63;
-    opacity: 0.6;
     position: absolute;
     display: inline-block;
     transition: all 0.3s ease;
+    opacity: 0.1;
 }
 
 .lnb li a:hover:after {
@@ -25,8 +25,8 @@
 }
 
 .lnb li.on a:after {
+    opacity: 0.3;
     width: 100%;
-    opacity: 1;
 }
 
 /* 내 근태 현황 */
@@ -510,4 +510,16 @@ table label {
 
 .fc-theme-standard .fc-popover {
     width: 14.66%;
+}
+
+.fc .fc-daygrid-day-frame {
+    min-height: 180px;
+}
+
+.fc-scroller {
+    overflow: hidden !important;
+}
+
+.fc-col-header, .fc-scrollgrid-sync-table, .fc-daygrid-body.fc-daygrid-body-unbalanced {
+    width: 100% !important;
 }
