+++ src/main/webapp/resources/js/approval.js
@@ -1,0 +1,6 @@
+$(function () {
+    // today default setting in modal
+    let todayDate = new Date().toISOString().substring(0, 10);
+    $('#apvFromDate').val(todayDate);
+    $('#apvToDate').val(todayDate);
+})
